refactor: plugins_block
