+++ packages/plugins/block/index.ts
@@ -11,8 +11,8 @@
  */
 
 import entry, { api } from './src/Main.vue'
-import metaData from './meta.js'
-import { BlockService } from './src/composable'
+import metaData from './meta'
+import { BlockService } from './src/composable/index'
 import SaveNewBlock from './src/SaveNewBlock.vue'
 import './src/styles/vars.less'
