+++ packages/plugins/block/src/BlockConfig.vue
@@ -107,7 +107,7 @@
   </tiny-form>
 </template>
 
-<script>
+<script lang="ts">
 import { reactive, ref, computed, nextTick, watchEffect } from 'vue'
 import { Input, Tag, Button, Form, FormItem, Radio, Select, Option } from '@opentiny/vue'
 import { constants } from '@opentiny/tiny-engine-utils'
