+++ packages/plugins/block/src/BlockGroupArrange.vue
@@ -11,7 +11,7 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
 import { reactive } from 'vue'
 
 export default {
