+++ packages/plugins/block/src/BlockPropertyForm.vue
@@ -112,7 +112,7 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
 import { computed, ref, watch } from 'vue'
 import {
   Input as TinyInput,
